+++ apps/web/src/components/BuilderNft/FaqItem.tsx
@@ -1,0 +1,8 @@
+export function FaqItem({ question, answer }: { question: string; answer: JSX.Element | string }) {
+  return (
+    <div className="flex flex-col gap-6 border-t border-translucent-200 pt-8">
+      <h2 className="text-2xl">{question}</h2>
+      <p>{answer}</p>
+    </div>
+  );
+}
